requirements.txt
typer>=0.9,<1.0
kubernetes>=28.1,<31.0
pydantic>=2.0,<3.0
networkx>=3.1,<4.0
openai>=1.0,<2.0
rich>=13.0,<14.0
PyYAML>=6.0,<7.0
